olm(channel): change from rhtas-1-0 to stable-1-0 (#359)
* olm(channel): change from rhtas-1-0 to stable-1-0

This commit modifies the channels we advertise for the operator,
removing the `rhtas-` prefix in favor of a `stable-` prefix, indicating
the major/minor version for that particular stable line. This channel
naming is based on the olm documentation best practices found at:
https://olm.operatorframework.io/docs/best-practices/channel-naming/#example-3-recommended-option-is-the-most-common-scenarios

I did not include a `fast-1-0` or `candidate-1-0`, however the current
configuration does not prevent these channels from being used in the
future. At this point, I don't think we have a strong enough cadence to
be using `fast` and `candidate`, however we may want to consider that
for the July release, or at least the EOY release.

If we want to follow the recommended semver naming conventions, then we
should consider the July release to be v1.1.0, given that it _should_
introduce new functionality with SECURESIGN-720 and possibly others
Jiras as well (key rotation, independently deployable services, etc).

If this PR is merged, the user would see only two channels when
installing `stable` and `stable-1-0`. They would (for now) contain the
exact same versions. But ultimately, as newer minor versions are
released, they will _all_ appear in `stable`, but `stable-1-0` will only
contain 1.0.0, 1.0.1, and any other 1.0.* releases we produce.

Do not merge this commit without discussion. I think we really want to get
it right and not have to make further changes.

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup

Signed-off-by: Lance Ball <lball@redhat.com>

---------

Signed-off-by: Lance Ball <lball@redhat.com>